+++ canvas_sdk/test_utils/factories/__init__.py
@@ -1,22 +1,49 @@
+from .calendar import CalendarFactory, EventFactory
 from .claim import (
     ClaimCommentFactory,
+    ClaimCoverageFactory,
     ClaimFactory,
     ClaimLabelFactory,
     ClaimProviderFactory,
     ClaimQueueFactory,
+    ClaimSubmissionFactory,
 )
 from .claim_diagnosis_code import ClaimDiagnosisCodeFactory
+from .coverage import CoverageFactory
 from .facility import FacilityFactory
+from .imaging import ImagingOrderFactory, ImagingReportFactory, ImagingReviewFactory
+from .lab import (
+    LabOrderFactory,
+    LabOrderReasonConditionFactory,
+    LabOrderReasonFactory,
+    LabPartnerFactory,
+    LabPartnerTestFactory,
+    LabReportFactory,
+    LabReviewFactory,
+    LabTestFactory,
+    LabValueCodingFactory,
+    LabValueFactory,
+)
 from .medication_history import (
     MedicationHistoryMedicationCodingFactory,
     MedicationHistoryMedicationFactory,
     MedicationHistoryResponseFactory,
 )
-from .note import NoteFactory, NoteStateChangeEventFactory
-from .organization import OrganizationFactory
+from .note import NoteFactory, NoteStateChangeEventFactory, NoteTypeFactory
+from .organization import (
+    OrganizationAddressFactory,
+    OrganizationContactPointFactory,
+    OrganizationFactory,
+)
 from .patient import PatientAddressFactory, PatientFacilityAddressFactory, PatientFactory
-from .practicelocation import PracticeLocationFactory, PracticeLocationSettingFactory
+from .practicelocation import (
+    PracticeLocationAddressFactory,
+    PracticeLocationContactPointFactory,
+    PracticeLocationFactory,
+    PracticeLocationSettingFactory,
+)
 from .protocol_current import ProtocolCurrentFactory
+from .referral import ReferralFactory, ReferralReportFactory, ReferralReviewFactory
 from .staff import (
     StaffAddressFactory,
     StaffContactPointFactory,
@@ -32,29 +59,59 @@ from .task import (
     TaskMetadataFactory,
     TaskTaskLabelFactory,
 )
+from .uncategorized_clinical_document import (
+    UncategorizedClinicalDocumentFactory,
+    UncategorizedClinicalDocumentReviewFactory,
+)
 from .user import CanvasUserFactory
 
 __all__ = (
+    "CalendarFactory",
     "CanvasUserFactory",
     "ClaimFactory",
     "ClaimCommentFactory",
+    "ClaimCoverageFactory",
     "ClaimDiagnosisCodeFactory",
     "ClaimLabelFactory",
     "ClaimProviderFactory",
     "ClaimQueueFactory",
+    "ClaimSubmissionFactory",
+    "CoverageFactory",
+    "EventFactory",
     "FacilityFactory",
+    "ImagingOrderFactory",
+    "ImagingReportFactory",
+    "ImagingReviewFactory",
+    "LabOrderFactory",
+    "LabOrderReasonConditionFactory",
+    "LabOrderReasonFactory",
+    "LabPartnerFactory",
+    "LabPartnerTestFactory",
+    "LabReportFactory",
+    "LabReviewFactory",
+    "LabTestFactory",
+    "LabValueCodingFactory",
+    "LabValueFactory",
     "MedicationHistoryMedicationFactory",
     "MedicationHistoryMedicationCodingFactory",
     "MedicationHistoryResponseFactory",
     "NoteFactory",
     "NoteStateChangeEventFactory",
+    "NoteTypeFactory",
+    "OrganizationAddressFactory",
+    "OrganizationContactPointFactory",
     "OrganizationFactory",
     "PatientAddressFactory",
     "PatientFacilityAddressFactory",
     "PatientFactory",
     "PracticeLocationFactory",
+    "PracticeLocationAddressFactory",
+    "PracticeLocationContactPointFactory",
     "PracticeLocationSettingFactory",
     "ProtocolCurrentFactory",
+    "ReferralFactory",
+    "ReferralReportFactory",
+    "ReferralReviewFactory",
     "StaffFactory",
     "StaffPhotoFactory",
     "StaffRoleFactory",
@@ -66,4 +123,6 @@ __all__ = (
     "TaskLabelFactory",
     "TaskMetadataFactory",
     "TaskTaskLabelFactory",
+    "UncategorizedClinicalDocumentFactory",
+    "UncategorizedClinicalDocumentReviewFactory",
 )