+++ plugin_runner/sandbox.py
@@ -90,6 +90,18 @@ STANDARD_LIBRARY_MODULES = {
         "Counter",
         "defaultdict",
     },
+    "dataclasses": {
+        "asdict",
+        "astuple",
+        "dataclass",
+        "field",
+        "Field",
+        "fields",
+        "InitVar",
+        "is_dataclass",
+        "make_dataclass",
+        "replace",
+    },
     "datetime": {
         "date",
         "datetime",
@@ -126,6 +138,7 @@ STANDARD_LIBRARY_MODULES = {
     "json": {
         "dumps",
         "loads",
+        "JSONDecodeError",
     },
     "operator": {
         "and_",
@@ -182,6 +195,9 @@ STANDARD_LIBRARY_MODULES = {
         "uuid4",
         "UUID",
     },
+    "defusedxml.ElementTree": {
+        "fromstring",
+    },
     "zoneinfo": {
         "ZoneInfo",
     },
@@ -195,19 +211,30 @@ THIRD_PARTY_MODULES = {
         "utcnow",
     },
     "django.db.models": {
+        "Avg",
         "BigIntegerField",
         "Case",
         "CharField",
         "Count",
+        "Exists",
         "F",
         "IntegerField",
+        "Max",
+        "Min",
         "Model",  # remove when hyperscribe no longer needs it
+        "OuterRef",
+        "Prefetch",
         "Q",
+        "Subquery",
+        "Sum",
         "Value",
         "When",
     },
     "django.db.models.expressions": {
         "Case",
+        "Exists",
+        "OuterRef",
+        "Subquery",
         "Value",
         "When",
     },
@@ -216,6 +243,7 @@ THIRD_PARTY_MODULES = {
         "Trim",
     },
     "django.db.models.query": {
+        "Prefetch",
         "QuerySet",
     },
     "django.utils.functional": {
@@ -663,6 +691,7 @@ class Sandbox:
                 "property": builtins.property,
                 "reversed": builtins.reversed,
                 "staticmethod": builtins.staticmethod,
+                "sum": builtins.sum,
                 "super": builtins.super,
                 "vars": builtins.vars,
             },